* Revert to old options and stock system (New changes are postponed to a future version)

+++ public_html/includes/entities/ent_attribute_group.inc.php
@@ -92,7 +92,7 @@
         }
 
         $product_option_values_query = database::query(
-          "select id from ". DB_TABLE_PREFIX ."products_customizations_values
+          "select id from ". DB_TABLE_PREFIX ."products_options_values
           where value_id = ". (int)$value['id'] ."
           limit 1;"
         );
@@ -170,6 +170,14 @@
 
         if (database::num_rows($products_attributes_query)) throw new Exception('Cannot delete value linked to product attributes');
 
+        $products_options_query = database::query(
+          "select id from ". DB_TABLE_PREFIX ."products_options_values
+          where value_id = ". (int)$value['id'] ."
+          limit 1;"
+        );
+
+        if (database::num_rows($products_options_query)) throw new Exception('Cannot delete value linked to product options');
+
         database::query(
           "delete from ". DB_TABLE_PREFIX ."attribute_values
           where group_id = ". (int)$this->data['id'] ."
@@ -198,7 +206,7 @@
 
         database::query(
           "update ". DB_TABLE_PREFIX ."attribute_values
-            priority = ". (int)$i++ .",
+          set priority = ". (int)$i++ .",
             date_updated = '". ($this->data['values'][$key]['date_updated'] = date('Y-m-d H:i:s')) ."'
           where id = ". (int)$value['id'] ."
           limit 1;"
@@ -242,9 +250,26 @@
 
       if (empty($this->data['id'])) return;
 
+    // Check products for attribute
+      $products_attributes_query = database::query(
+        "select id from ". DB_TABLE_PREFIX ."products_attributes
+        where group_id = ". (int)$this->data['id'] .";"
+      );
+
+      if (database::num_rows($products_attributes_query)) throw new Exception('Cannot delete group linked to products');
+
+    // Check products for options
+      $products_options_query = database::query(
+        "select id from ". DB_TABLE_PREFIX ."products_attributes
+        where group_id = ". (int)$this->data['id'] .";"
+      );
+
+      if (database::num_rows($products_options_query)) throw new Exception('Cannot delete group linked to products');
+
       $this->data['values'] = [];
       $this->save();
 
+    // Delete attribute
       database::query(
         "delete ag, agi, av, avi
         from ". DB_TABLE_PREFIX ."attribute_groups ag
